Move config widgets to the widgets module
That's because we plan to have other widgets besides plugin preference
pages that rely on them.
