+++ spyder/widgets/config.py
